+++ packages/reporting/test/index.js
@@ -14,8 +14,8 @@ import './url-analyzer.spec.js';
 import './alive-check.spec.js';
 import './sanitizer.spec.js';
 import './search-extractor.spec.js';
-import './request/unit/utils-test.spec.js';
-import './request/unit/step-context-test.spec.js';
-import './request/unit/qs-whitelist2-test.spec.js';
-import './request/unit/hash-test.spec.js';
-import './request/unit/index.spec.js';
+import './request/utils.spec.js';
+import './request/step-context.spec.js';
+import './request/qs-whitelist2.spec.js';
+import './request/hash.spec.js';
+import './request/index.spec.js';
